# Improve Fused8BitRowwiseQuantizedSBFloatToFloatOrHalfNeon by 2%-10%